requirements.txt
adal==1.2.1
aiounittest==1.3.0
alabaster==0.7.12
appnope==0.1.0
asgiref==3.3.1
asn1crypto==1.3.0
astroid==2.3.3
asyncio==3.4.3
attrs==19.3.0
autopep8==1.5.4
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
botbuilder-core==4.11.0
botbuilder-dialogs==4.11.0
botbuilder-schema==4.11.0
botframework-connector==4.11.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.4
cryptography==2.8
datedelta==1.3
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.3
django-admin-relation-links==0.2.5
django-dynamic-forms==0.5.3
django-pwa==1.0.10
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
docutils==0.16
emoji==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
grapheme==0.6.0
gunicorn==20.0.4
h11==0.11.0
httplib2==0.19.0
httptools==0.1.1
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
ippanel==1.1.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.21
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
jsonfield==3.1.0
jsonpickle==1.2
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
keyring==21.1.0
lazy-object-proxy==1.4.3
mailchimp-transactional==1.0.23
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
msal==1.2.0
msrest==0.6.10
multipledispatch==0.6.0
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
numpydoc==0.9.2
oauthlib==3.1.0
openpyxl==3.0.6
packaging==20.1
pandas~=1.2.3
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prompt-toolkit==3.0.3
psutil==5.6.7
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.19
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.6
PyQt5==5.12.3
PyQt5-sip==12.8.1
PyQtWebEngine==5.12.1
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.15.0
python-http-client==3.3.2
pytz==2019.3
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==5.3.1
pyzmq==18.1.1
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
# recognizers-text==1.0.2a2
# recognizers-text-choice==1.0.2a2
# recognizers-text-date-time==1.0.2a2
# recognizers-text-number==1.0.2a2
# recognizers-text-number-with-unit==1.0.2a2
regex==2019.8.19
requests==2.23.0
requests-oauthlib==1.3.0
rope==0.16.0
ruamel.yaml==0.15.87
sendgrid==6.4.8
six==1.14.0
snowballstemmer==2.0.0
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.3.6
spyder-kernels==0.5.2
sqlparse==0.3.0
starkbank-ecdsa==1.1.0
testpath==0.4.4
toml==0.10.2
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
twilio==6.53.0
typing-extensions==3.7.4.3
Unipath==1.1
urllib3==1.25.8
uvicorn==0.12.2
watchgod==0.6
wcwidth==0.1.8
webencodings==0.5.1
websockets==8.1
whitenoise==5.2.0
wrapt==1.11.2
wurlitzer==2.0.0
zipp==2.2.0
django-timedeltafield~=0.7.10